TEZ-1116. Refactor YarnTezDAGChild to be testable and usable for LocalMode. (sseth)

+++ tez-api/src/main/java/org/apache/tez/dag/api/TezConfiguration.java
@@ -180,7 +180,7 @@ public class TezConfiguration extends Configuration {
    */
   public static final String TEZ_TASK_AM_HEARTBEAT_COUNTER_INTERVAL_MS = TEZ_TASK_PREFIX
       + "am.heartbeat.counter.interval-ms.max";
-  public static final long TEZ_TASK_AM_HEARTBEAT_COUNTER_INTERVAL_MS_DEFAULT =
+  public static final int TEZ_TASK_AM_HEARTBEAT_COUNTER_INTERVAL_MS_DEFAULT =
       1000;
 
   public static final String TEZ_TASK_MAX_EVENTS_PER_HEARTBEAT = TEZ_TASK_PREFIX
